refactor: NewsViewer 분리
